requirements.txt
ImageIO==2.37.2
lazy_loader==0.4
networkx==3.2.1
numpy==2.0.2
opencv-python==4.13.0.90
packaging==26.0
pillow==11.3.0
scikit-image==0.24.0
scipy==1.13.1
tifffile==2024.8.30
